pyproject.toml
[project]
name = "poke-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "cachetools>=6.0.0",
    "flagembedding>=1.3.5",
    "lxml>=6.1.0",  # CVE-2026-41066 patched in 6.1.0; transitive via flagembedding→ir-datasets
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.17.1",
    "peft>=0.14.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.0,<6.0.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.44.0",
]
dev = [
    "anyio>=4.0.0",
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.11",
]
cache = [
    "cachetools>=6.0.0",
    "redis[hiredis]>=5.0.0",
]
train = [
    # RunPod / CUDA only — do not install locally
    "unsloth",
    "trl",
    "bitsandbytes",
    "datasets",
]
gen = [
    "google-genai>=1.0.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "unit: no I/O, no model",
    "integration: real I/O against fixture data",
    "e2e: full pipeline with real model",
    "gpu: requires a CUDA device",
    "slow: takes more than 5 seconds",
    "anyio: run test with anyio async backend",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.11"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = ["cachetools", "redis.*"]
ignore_missing_imports = true
